Refactoring pass 1 complete -- moving on to adding in Box ID naming convention changes
