+++ Json Validator/JsonPopulator/JsonPopulator.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsonPopulator", "JsonPopula
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JsonValidator", "JsonValidator\JsonValidator.csproj", "{2398200C-5D34-4674-B0FA-4877CE17659E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PopOwnership", "PopOwnership\PopOwnership.csproj", "{D0775812-29D4-44A0-9A66-7C377CB84B90}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -23,10 +21,6 @@ Global
 		{2398200C-5D34-4674-B0FA-4877CE17659E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2398200C-5D34-4674-B0FA-4877CE17659E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2398200C-5D34-4674-B0FA-4877CE17659E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D0775812-29D4-44A0-9A66-7C377CB84B90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D0775812-29D4-44A0-9A66-7C377CB84B90}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D0775812-29D4-44A0-9A66-7C377CB84B90}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D0775812-29D4-44A0-9A66-7C377CB84B90}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
