+++ Json Validator/JsonPopulator/JsonValidator/JsonControllers/JsonGeneration.cs
@@ -11,7 +11,8 @@ using JsonValidator.JsonConverters;
 namespace JsonValidator
 {
     public class JsonGeneration
-    {
+    { 
+        //all members for generating the Json File
         public void GenerateMyJson(Form1 form)
         {
             var comboBoxes = form.Controls.OfType<ComboBox>().ToList();
@@ -19,16 +20,18 @@ namespace JsonValidator
             var dateTimePicker = form.Controls.OfType<DateTimePicker>().ToList();
             var checkBoxes = form.Controls.OfType<CheckBox>().ToList();
             var textBoxes = form.Controls.OfType<TextBox>().ToList();
-
-            AppearanceConverter appearance = new AppearanceConverter(form);
+            
+            //classes for assigning form values to NewRoot object properties
+            AppearanceConverter appearance = new AppearanceConverter(form); 
             PrizesConverter prizesConverter = new PrizesConverter();
             TierConverter tierConverter = new TierConverter();
+
 
             NewRoot finalRoot = new NewRoot()
             {
                 boxId = comboBoxes.Find(x => x.Name == "boxIdCB").Text,
                 fandomId = textBoxes.Find(x => x.Name == "fandomIdCB").Text,
-                startDate = dateTimePicker.Find(x => x.Name == "startDatePicker").Text,
+                startDate =  dateTimePicker.Find(x => x.Name == "startDatePicker").Text,
                 endDate = dateTimePicker.Find(x => x.Name == "endDatePicker").Text,
                 appearance = appearance.GenerateAppearance(),
                 behaviourType = comboBoxes.Find(x => x.Name == "behaviorCB").Text,
@@ -38,11 +41,11 @@ namespace JsonValidator
                 lastChanceBoxPrizes = GetLastChanceList(flowBoxes),
             };
 
-            Debug.WriteLine(FormatJson(finalRoot));
-            File.WriteAllText(@"C:\Users\pdnud\OneDrive\Desktop\Json Validator\[1.5.0] mystery_boxes_config - Default.json", FormatJson(finalRoot));
+            finalRoot.FixDates(finalRoot.startDate, finalRoot.endDate);
+
+            File.WriteAllText(@"C:\Users\pdnud\OneDrive\Desktop\Json Validator\[1.5.0] mystery_boxes_config - Default.json", SerializeJson(finalRoot));
             System.Diagnostics.Process.Start(@"C:\Users\pdnud\OneDrive\Desktop\Json Validator\[1.5.0] mystery_boxes_config - Default.json");
         }
-
         public List<LastChanceBoxPrize> GetLastChanceList(List<FlowLayoutPanel> flowlist)
         {
             var popLists = flowlist.Find(x => x.Name == "lastChanceBoxPanel").Controls.OfType<GroupBox>().ToList();
@@ -58,7 +61,6 @@ namespace JsonValidator
                 lastCBox.instances = Convert.ToInt32(combos[3].Text);
                 lastChanceList.Add(lastCBox);
             }
-
             return lastChanceList;
         }
         public List<string> GetFeaturedPops(List<FlowLayoutPanel> flowList)
@@ -66,27 +68,13 @@ namespace JsonValidator
             var popLists = flowList.Find(x => x.Name == "storePopsPanel").Controls.OfType<ComboBox>().ToList();
             List<string> _popIds = new List<string>();
 
-            foreach (var x in popLists) //break me into a new class with store appearance
+            foreach (var x in popLists)
             {
                 _popIds.Add(x.Text);
             }
             return _popIds;
         }
-
         public string SerializeJson(NewRoot newRoot)
-        {
-            //Accepts Completed NewRoot Object
-            var serializerSettings = new JsonSerializerSettings();
-
-            string rootOutput = JsonConvert.SerializeObject(newRoot, Formatting.Indented, new JsonSerializerSettings
-            {
-                NullValueHandling = NullValueHandling.Ignore
-            });
-
-            return rootOutput;
-        }
-
-        public string FormatJson(NewRoot newRoot)
         {
             var settings = new JsonSerializerSettings
             {
@@ -98,27 +86,5 @@ namespace JsonValidator
             string json = JsonConvert.SerializeObject(newRoot, settings);
             return json;
         }
-
     }
-    class MyConverter : JsonConverter
-    {
-        public override bool CanConvert(Type objectType) //controlling what types are affected with formatting, indented or none
-        {
-            return objectType == typeof(string[]) || objectType == typeof(Tier) || objectType == typeof(Prize) 
-                || objectType == typeof(List<string>) || objectType == typeof(LastChanceBoxPrize);
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            
-            writer.WriteRawValue(JsonConvert.SerializeObject(value, Formatting.None, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore }));
-        }
-
-    }
-
 }
