+++ Json Validator/JsonPopulator/JsonValidator/JsonConverters/AppearanceConverter.cs
@@ -25,14 +25,18 @@ namespace JsonValidator.JsonConverters
             var comboBoxes = form.Controls.OfType<ComboBox>().ToList();
             var checkBoxes = form.Controls.OfType<CheckBox>().ToList();
 
+
             Appearance appearance = new Appearance()
             {
                 isEventBox = checkBoxes.Find(x => x.Name == "isEventCheck").Checked,
                 mysteryBoxType = comboBoxes.Find(x => x.Name == "MysteryBoxCB").Text,
+                isOEDBox = checkBoxes.Find(x => x.Name == "oedBoxCheck").Checked,
+                isVIPBox = checkBoxes.Find(x => x.Name == "isVIPBoxCheck").Checked,
+
                 theme = comboBoxes.Find(x => x.Name == "themeCB").Text,
                 storeButtonAppearance = sBA.GenerateStoreBA(),
                 purchaseScreenAppearance = pSA.GeneratePurchaseScreenApeparance(),
-                mainHubAppearance = mHA.GenerateHubApp()
+                mainHubAppearance = mHA.GenerateHubApp(),
             };
 
             return appearance;
