+++ Json Validator/JsonPopulator/JsonValidator/JsonValidator.csproj
@@ -83,6 +83,7 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
+    <Compile Include="JsonControllers\FormControls.cs" />
     <Compile Include="FormObjects\ITierBox.cs" />
     <Compile Include="Json Classes\Appearance.cs" />
     <Compile Include="JsonConverters\AppearanceConverter.cs" />
@@ -96,12 +97,14 @@
     <Compile Include="Json Classes\StoreButtonAppearance.cs" />
     <Compile Include="Json Classes\Tier.cs" />
     <Compile Include="FormObjects\PrizeBox.cs" />
-    <Compile Include="JsonConverters\JsonGeneration.cs" />
+    <Compile Include="JsonControllers\JsonGeneration.cs" />
     <Compile Include="JsonConverters\MainHubConverter.cs" />
+    <Compile Include="JsonConverters\MyConverter.cs" />
     <Compile Include="JsonConverters\PrizesConverter.cs" />
     <Compile Include="JsonConverters\PurchaseScreenConverter.cs" />
     <Compile Include="JsonConverters\StoreButtonConverter.cs" />
     <Compile Include="JsonConverters\TierConverter.cs" />
+    <Compile Include="Miscellaneous\Misc.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="FormObjects\TierBoxL.cs" />
